refactor: purify phosphor-geometry, move settings/QScreen deps to proper layers
phosphor-geometry is now Qt6::Core only with zero interfaces:
- Remove IGeometrySettings (restored in engine-api where it belongs)
- Remove QScreen* overload (moved to engine-api + phosphor-zones)
- Move zone-assignment JsonKeys to engine-api (domain, not geometry)

phosphor-zones API is now settings-agnostic:
- getZoneGeometryForScreen[F] takes resolved int/EdgeGaps values
- Gap resolution cascade moved to daemon (src/core/geometryutils.cpp)

Also: fix doc branch name, clarify PR numbering, add namespace
note to engine-types CMakeLists.

All 131 tests pass. Zero behavioral changes.
