+++ libs/phosphor-engine-api/CMakeLists.txt
@@ -52,6 +52,7 @@ target_compile_features(PhosphorEngineApi PUBLIC cxx_std_20)
 target_link_libraries(PhosphorEngineApi
     PUBLIC
         Qt6::Core
+        Qt6::Gui
         PhosphorEngineTypes::PhosphorEngineTypes
         PhosphorGeometry::PhosphorGeometry
 )
