+++ libs/phosphor-engine-api/include/PhosphorEngineApi/JsonKeys.h
@@ -5,15 +5,20 @@
 
 #include <PhosphorGeometry/JsonKeys.h>
 
+#include <QLatin1String>
+
 namespace PhosphorEngineApi {
 namespace JsonKeys {
+
 using PhosphorGeometry::JsonKeys::Height;
-using PhosphorGeometry::JsonKeys::SourceZoneId;
-using PhosphorGeometry::JsonKeys::TargetZoneId;
-using PhosphorGeometry::JsonKeys::TargetZoneIds;
 using PhosphorGeometry::JsonKeys::Width;
-using PhosphorGeometry::JsonKeys::WindowId;
 using PhosphorGeometry::JsonKeys::X;
 using PhosphorGeometry::JsonKeys::Y;
+
+inline constexpr QLatin1String WindowId{"windowId"};
+inline constexpr QLatin1String SourceZoneId{"sourceZoneId"};
+inline constexpr QLatin1String TargetZoneId{"targetZoneId"};
+inline constexpr QLatin1String TargetZoneIds{"targetZoneIds"};
+
 } // namespace JsonKeys
 } // namespace PhosphorEngineApi
