+++ libs/phosphor-geometry/PhosphorGeometryConfig.cmake.in
@@ -5,7 +5,7 @@
 
 include(CMakeFindDependencyMacro)
 
-find_dependency(Qt6 6.6 COMPONENTS Core Gui)
+find_dependency(Qt6 6.6 COMPONENTS Core)
 
 include("${CMAKE_CURRENT_LIST_DIR}/PhosphorGeometryTargets.cmake")
