+++ libs/phosphor-geometry/include/PhosphorGeometry/JsonKeys.h
@@ -13,10 +13,5 @@ inline constexpr QLatin1String Y{"y"};
 inline constexpr QLatin1String Width{"width"};
 inline constexpr QLatin1String Height{"height"};
 
-inline constexpr QLatin1String WindowId{"windowId"};
-inline constexpr QLatin1String SourceZoneId{"sourceZoneId"};
-inline constexpr QLatin1String TargetZoneId{"targetZoneId"};
-inline constexpr QLatin1String TargetZoneIds{"targetZoneIds"};
-
 } // namespace JsonKeys
 } // namespace PhosphorGeometry
