+++ libs/phosphor-zones/include/PhosphorZones/GeometryUtils.h
@@ -4,7 +4,6 @@
 #pragma once
 
 #include <PhosphorGeometry/GeometryUtils.h>
-#include <PhosphorGeometry/IGeometrySettings.h>
 #include <phosphorzones_export.h>
 
 #include <PhosphorLayoutApi/EdgeGaps.h>
@@ -34,6 +33,8 @@ using PhosphorGeometry::rectToJson;
 using PhosphorGeometry::removeZoneOverlaps;
 using PhosphorGeometry::snapToRect;
 
+PHOSPHORZONES_EXPORT QRectF availableAreaToOverlayCoordinates(const QRectF& geometry, QScreen* screen);
+
 PHOSPHORZONES_EXPORT QRectF getZoneGeometryWithGaps(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone,
                                                     QScreen* screen, int innerGap,
                                                     const ::PhosphorLayout::EdgeGaps& outerGaps,
@@ -46,21 +47,13 @@ PHOSPHORZONES_EXPORT QRectF getZoneGeometryWithGaps(Phosphor::Screens::ScreenMan
 
 PHOSPHORZONES_EXPORT QRect getZoneGeometryForScreen(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone,
                                                     QScreen* screen, const QString& screenId,
-                                                    PhosphorZones::Layout* layout,
-                                                    PhosphorGeometry::IGeometrySettings* settings);
+                                                    PhosphorZones::Layout* layout, int zonePadding,
+                                                    const ::PhosphorLayout::EdgeGaps& outerGaps);
 
 PHOSPHORZONES_EXPORT QRectF getZoneGeometryForScreenF(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone,
                                                       QScreen* screen, const QString& screenId,
-                                                      PhosphorZones::Layout* layout,
-                                                      PhosphorGeometry::IGeometrySettings* settings);
-
-PHOSPHORZONES_EXPORT int getEffectiveZonePadding(PhosphorZones::Layout* layout,
-                                                 PhosphorGeometry::IGeometrySettings* settings,
-                                                 const QString& screenId = {});
-
-PHOSPHORZONES_EXPORT ::PhosphorLayout::EdgeGaps getEffectiveOuterGaps(PhosphorZones::Layout* layout,
-                                                                      PhosphorGeometry::IGeometrySettings* settings,
-                                                                      const QString& screenId = {});
+                                                      PhosphorZones::Layout* layout, int zonePadding,
+                                                      const ::PhosphorLayout::EdgeGaps& outerGaps);
 
 PHOSPHORZONES_EXPORT QRectF effectiveScreenGeometry(Phosphor::Screens::ScreenManager* mgr,
                                                     PhosphorZones::Layout* layout, QScreen* screen);
