+++ src/core/geometryutils.cpp
@@ -8,6 +8,7 @@
 #include "interfaces.h"
 #include "constants.h"
 #include <PhosphorEngineApi/IGeometrySettings.h>
+#include <PhosphorIdentity/VirtualScreenId.h>
 #include <PhosphorScreens/Manager.h>
 #include <PhosphorScreens/ScreenIdentity.h>
 #include <PhosphorZones/ZoneDefaults.h>
@@ -36,29 +37,118 @@ ScreenGeometries resolveScreenGeometries(Phosphor::Screens::ScreenManager* mgr, 
     }
     return {mgr->screenGeometry(screenId), mgr->screenAvailableGeometry(screenId)};
 }
+
+QVariantMap getPerScreenSnappingWithFallback(PhosphorEngineApi::IGeometrySettings* settings, const QString& screenId)
+{
+    QVariantMap result = settings->getPerScreenSnappingSettings(screenId);
+    if (result.isEmpty() && PhosphorIdentity::VirtualScreenId::isVirtual(screenId)) {
+        result = settings->getPerScreenSnappingSettings(PhosphorIdentity::VirtualScreenId::extractPhysicalId(screenId));
+    }
+    return result;
+}
 } // anonymous namespace
-
-QRectF getZoneGeometryForScreenF(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone, QScreen* screen,
-                                 const QString& screenId, PhosphorZones::Layout* layout, ISettings* settings)
-{
-    return ::PhosphorZones::GeometryUtils::getZoneGeometryForScreenF(mgr, zone, screen, screenId, layout, settings);
-}
-
-QRect getZoneGeometryForScreen(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone, QScreen* screen,
-                               const QString& screenId, PhosphorZones::Layout* layout, ISettings* settings)
-{
-    return ::PhosphorZones::GeometryUtils::getZoneGeometryForScreen(mgr, zone, screen, screenId, layout, settings);
-}
 
 int getEffectiveZonePadding(PhosphorZones::Layout* layout, ISettings* settings, const QString& screenId)
 {
-    return ::PhosphorZones::GeometryUtils::getEffectiveZonePadding(layout, settings, screenId);
+    namespace PSK = PhosphorEngineApi::PerScreenSnappingKey;
+    if (!screenId.isEmpty() && settings) {
+        QVariantMap perScreen = getPerScreenSnappingWithFallback(settings, screenId);
+        auto it = perScreen.constFind(PSK::ZonePadding);
+        if (it != perScreen.constEnd()) {
+            return it->toInt();
+        }
+    }
+    if (layout && layout->hasZonePaddingOverride()) {
+        return layout->zonePadding();
+    }
+    if (settings) {
+        return settings->zonePadding();
+    }
+    return PhosphorEngineApi::GeometryDefaults::ZonePadding;
 }
 
 ::PhosphorLayout::EdgeGaps getEffectiveOuterGaps(PhosphorZones::Layout* layout, ISettings* settings,
                                                  const QString& screenId)
 {
-    return ::PhosphorZones::GeometryUtils::getEffectiveOuterGaps(layout, settings, screenId);
+    namespace PSK = PhosphorEngineApi::PerScreenSnappingKey;
+    namespace GD = PhosphorEngineApi::GeometryDefaults;
+    if (!screenId.isEmpty() && settings) {
+        QVariantMap perScreen = getPerScreenSnappingWithFallback(settings, screenId);
+        if (!perScreen.isEmpty()) {
+            auto usePerSideIt = perScreen.constFind(PSK::UsePerSideOuterGap);
+            bool usePerSide = (usePerSideIt != perScreen.constEnd()) ? usePerSideIt->toBool() : false;
+            if (usePerSide) {
+                auto topIt = perScreen.constFind(PSK::OuterGapTop);
+                auto bottomIt = perScreen.constFind(PSK::OuterGapBottom);
+                auto leftIt = perScreen.constFind(PSK::OuterGapLeft);
+                auto rightIt = perScreen.constFind(PSK::OuterGapRight);
+                if (topIt != perScreen.constEnd() || bottomIt != perScreen.constEnd() || leftIt != perScreen.constEnd()
+                    || rightIt != perScreen.constEnd()) {
+                    auto uniformIt = perScreen.constFind(PSK::OuterGap);
+                    int fallback = (uniformIt != perScreen.constEnd()) ? uniformIt->toInt() : settings->outerGap();
+                    return {(topIt != perScreen.constEnd()) ? topIt->toInt() : fallback,
+                            (bottomIt != perScreen.constEnd()) ? bottomIt->toInt() : fallback,
+                            (leftIt != perScreen.constEnd()) ? leftIt->toInt() : fallback,
+                            (rightIt != perScreen.constEnd()) ? rightIt->toInt() : fallback};
+                }
+            }
+            auto uniformIt = perScreen.constFind(PSK::OuterGap);
+            if (uniformIt != perScreen.constEnd()) {
+                return ::PhosphorLayout::EdgeGaps::uniform(uniformIt->toInt());
+            }
+        }
+    }
+    if (layout && layout->usePerSideOuterGap() && layout->hasPerSideOuterGapOverride()) {
+        ::PhosphorLayout::EdgeGaps gaps = layout->rawOuterGaps();
+        if (settings && settings->usePerSideOuterGap()) {
+            if (gaps.top < 0)
+                gaps.top = settings->outerGapTop();
+            if (gaps.bottom < 0)
+                gaps.bottom = settings->outerGapBottom();
+            if (gaps.left < 0)
+                gaps.left = settings->outerGapLeft();
+            if (gaps.right < 0)
+                gaps.right = settings->outerGapRight();
+        } else {
+            int fallback = settings ? settings->outerGap() : GD::OuterGap;
+            if (gaps.top < 0)
+                gaps.top = fallback;
+            if (gaps.bottom < 0)
+                gaps.bottom = fallback;
+            if (gaps.left < 0)
+                gaps.left = fallback;
+            if (gaps.right < 0)
+                gaps.right = fallback;
+        }
+        return gaps;
+    }
+    if (layout && layout->hasOuterGapOverride()) {
+        return ::PhosphorLayout::EdgeGaps::uniform(layout->outerGap());
+    }
+    if (settings) {
+        if (settings->usePerSideOuterGap()) {
+            return {settings->outerGapTop(), settings->outerGapBottom(), settings->outerGapLeft(),
+                    settings->outerGapRight()};
+        }
+        return ::PhosphorLayout::EdgeGaps::uniform(settings->outerGap());
+    }
+    return ::PhosphorLayout::EdgeGaps::uniform(GD::OuterGap);
+}
+
+QRectF getZoneGeometryForScreenF(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone, QScreen* screen,
+                                 const QString& screenId, PhosphorZones::Layout* layout, ISettings* settings)
+{
+    int zp = getEffectiveZonePadding(layout, settings, screenId);
+    ::PhosphorLayout::EdgeGaps og = getEffectiveOuterGaps(layout, settings, screenId);
+    return ::PhosphorZones::GeometryUtils::getZoneGeometryForScreenF(mgr, zone, screen, screenId, layout, zp, og);
+}
+
+QRect getZoneGeometryForScreen(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone, QScreen* screen,
+                               const QString& screenId, PhosphorZones::Layout* layout, ISettings* settings)
+{
+    int zp = getEffectiveZonePadding(layout, settings, screenId);
+    ::PhosphorLayout::EdgeGaps og = getEffectiveOuterGaps(layout, settings, screenId);
+    return ::PhosphorZones::GeometryUtils::getZoneGeometryForScreen(mgr, zone, screen, screenId, layout, zp, og);
 }
 
 static EmptyZoneList buildEmptyZoneListImpl(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Layout* layout,
